More flexibility for external code analysis tool
Instead of assuming that there is a static code analyzer with a fixed
command line interface, allow to specify an arbitrary command line
for a code checker tool.
Also assume that diagnostics output JSON is written to an external
temporary file (more flexibility again).
This is needed to be able to migrate to a new analyzer and allow to
switching between different tools which can include project-specific
bindings and other features.